fix: cleaned up code
