+++ crates/common/consensus/src/execution_engine/rpc_types/mod.rs
@@ -1,3 +1,4 @@
-pub mod engine_get_payloadv3;
-pub mod engine_new_payloadv3;
 pub mod eth_syncing;
+pub mod execution_payload;
+pub mod get_payload;
+pub mod payload_status;
